// src/app/common-library/common-types/common-type.tsx
export interface PaginationProps {
  limit?: number;
  page?: number;
  sortType: string;
  sortBy: string;
}


export interface SortProps {
  [t: string]: 'asc' | 'des';
}
// ...
export type GetAllPropsServer<T> = ({
                                      queryProps,
                                      sortList,
                                      paginationProps,
                                    }: {
  queryProps: any;
  sortList?: SortProps[];
  paginationProps?: PaginationProps;
}) => (Promise<AxiosResponse<{ data: T[], paging: any }>>);
// ...
export type CountProps<T> = ({
                               queryProps,
                               sortList,
                               paginationProps,
                             }: {
  queryProps: any;
  sortList?: SortProps[];
  paginationProps?: PaginationProps;
}) => Promise<AxiosResponse>;
export type RenderInfoDetail = {
  header?: string;
  className?: string;
  titleClassName?: string;
  dataClassName?: string;
  style?: string;
  data: RenderInfoDetailColumn,
}[]
export type RenderInfoDetailColumn = {
  [T: string]: {
    title?: string;
    formatter?: (value: any | any[], entity?: any) => ReactElement;
    keyField?: string;
  }
}
export type GetProps<T> = (entity: T) => Promise<AxiosResponse>;
export type UpdateProps<T> = (entity: T) => Promise<AxiosResponse>;
export type ApproveProps<T> = (entity: T, data: T) => Promise<AxiosResponse>;
export type DeleteProps<T> = (entity: T) => Promise<AxiosResponse>;
export type CreateProps<T> = (entity: T) => Promise<AxiosResponse>;
export type DeleteManyProps<T> = (entities: T[]) => Promise<AxiosResponse>;